change from default IO to VFS

+++ core/src/util/data_structures/data_types.h
@@ -5,24 +5,25 @@
 
 // TYPES ===========================================================================================================
 
-typedef uint8_t 				u8;		// 8-bit unsigned integer
-typedef uint16_t 				u16;	// 16-bit unsigned integer
-typedef uint32_t 				u32;	// 32-bit unsigned integer
-typedef uint64_t 				u64;	// 64-bit unsigned integer
+typedef uint_fast8_t 				u8;		// 8-bit unsigned integer
+typedef uint_fast16_t 				u16;	// 16-bit unsigned integer
+typedef uint_fast32_t 				u32;	// 32-bit unsigned integer
+typedef uint_fast64_t 				u64;	// 64-bit unsigned integer
 
-typedef int8_t 					i8;	 	// 8-bit signed integer
-typedef int16_t 				i16; 	// 16-bit signed integer
-typedef int32_t 				i32; 	// 32-bit signed integer
-typedef int64_t 				i64; 	// 64-bit signed integer
+typedef int_fast8_t 				i8;	 	// 8-bit signed integer
+typedef int_fast16_t 				i16; 	// 16-bit signed integer
+typedef int_fast32_t 				i32; 	// 32-bit signed integer
+typedef int_fast64_t 				i64; 	// 64-bit signed integer
 
-typedef float 					f32;	// 32-bit floating point
-typedef double 					f64;	// 64-bit floating point
-typedef long double 			f128;	// 128-bit floating point (platform dependent)
+typedef float 						f32;	// 32-bit floating point
+typedef double 						f64;	// 64-bit floating point
+typedef long double 				f128;	// 128-bit floating point (platform dependent)
 
 // Platform-specific types
-typedef unsigned long long 		handle; // Generic handle type for OS resources
+typedef u64 						handle; // Generic handle type for OS resources
 
 // FORWARD DECLARATIONS ================================================================================================
+
 
 namespace GLT {
